(fixup) Move TestUtils to testing module so that every module can utilize it
