+++ .travis.yml
@@ -13,3 +13,5 @@ cache:
   directories:
     - $HOME/.gradle/caches/
     - $HOME/.gradle/wrapper/
+script:
+  - ./gradlew build
