+++ build.gradle
@@ -210,6 +210,6 @@ subprojects {
             archiveClassifier = 'javadoc'
             from javadoc
         }
-        check.dependsOn(javadoc, integrationTest)
+        build.dependsOn(javadoc, integrationTest)
     }
 }
