+++ testing/build.gradle
@@ -18,6 +18,9 @@ ext {
     noPublish = true
 }
 dependencies {
+    implementation project(':processor')
+    implementation project(':protobuf')
+
     implementation "org.apache.kafka:kafka-clients:$kafkaVersion"
     implementation "org.apache.kafka:kafka_2.12:$kafkaVersion"
     implementation "org.apache.kafka:kafka-clients:$kafkaVersion:test"
