+++ testing/src/main/java/com/linecorp/decaton/testing/TestUtils.java
@@ -14,16 +14,18 @@
  * under the License.
  */
 
-package com.linecorp.decaton.processor;
+package com.linecorp.decaton.testing;
 
 import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.kafka.clients.producer.ProducerConfig;
 
 import com.google.protobuf.MessageLite;
 
 import com.linecorp.decaton.client.DecatonClient;
+import com.linecorp.decaton.processor.ProcessorsBuilder;
 import com.linecorp.decaton.processor.SubscriptionStateListener.State;
 import com.linecorp.decaton.processor.runtime.ProcessorSubscription;
 import com.linecorp.decaton.processor.runtime.RetryConfig;
@@ -31,12 +33,18 @@ import com.linecorp.decaton.processor.runtime.SubscriptionBuilder;
 import com.linecorp.decaton.protobuf.ProtocolBuffersSerializer;
 
 public class TestUtils {
+    private static final AtomicInteger sequence = new AtomicInteger(0);
+
+    // generate a monotonic value to be added as part of client.id to ensure unique
+    private static int sequence() {
+        return sequence.getAndIncrement();
+    }
 
     public static <T extends MessageLite> DecatonClient<T> client(String topic,
                                                                   String bootstrapServers) {
         Properties props = new Properties();
         props.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        props.setProperty(ProducerConfig.CLIENT_ID_CONFIG, "test-client");
+        props.setProperty(ProducerConfig.CLIENT_ID_CONFIG, "test-client-" + sequence());
         return DecatonClient.producing(topic, new ProtocolBuffersSerializer<T>())
                             .applicationId("test-application")
                             .instanceId("test-instance")
@@ -49,7 +57,7 @@ public class TestUtils {
                                                          RetryConfig retryConfig) {
         Properties props = new Properties();
         props.setProperty("bootstrap.servers", bootstrapServers);
-        props.setProperty("client.id", "test-processor");
+        props.setProperty("client.id", "test-processor-" + sequence());
         props.setProperty("group.id", "test-group");
         CountDownLatch initializationLatch = new CountDownLatch(1);
         ProcessorSubscription subscription = SubscriptionBuilder.newBuilder("test-subscription")
